seperate arcade & token loader to separate comp
