+++ ui/src/app/components/navigation/Header.tsx
@@ -122,7 +122,7 @@ export default function Header({
           size={"xs"}
           variant={checkArcade ? "outline" : "default"}
           onClick={() => showArcadeDialog(!arcadeDialog)}
-          disabled={isWrongNetwork}
+          disabled={isWrongNetwork || !account}
           className={`xl:px-5 ${checkArcade ? "" : "animate-pulse"}`}
         >
           <ArcadeIcon className="sm:w-5 sm:h-5  h-3 w-3 justify-center fill-current mr-2" />
